interview/src/test/java/apple/_1547_Minimum_Cost_to_Cut_a_Stick/SolutionITest.java
package apple._1547_Minimum_Cost_to_Cut_a_Stick;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 3/2/2021
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final int n = 7, cuts[] = {1, 3, 4, 5}, expected = 16, actual = solution.minCost(n, cuts);
    assertEquals(expected, actual);
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase2(Solution solution) {
    final int n = 9, cuts[] = {5, 6, 1, 4, 2}, expected = 22, actual = solution.minCost(n, cuts);
    assertEquals(expected, actual);
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}